requirements.txt
certifi==2019.6.16
chardet==3.0.4
docopt==0.6.1
future==0.17.1
httplib2==0.13.0
httpretty==0.9.6
idna==2.8
oauthlib==3.0.1
PyTumblr==0.0.8
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.3
